# Refs #32568 -- Optimized escape() by using SafeString instead of mark_safe().